+++ changelog.md
@@ -3,6 +3,7 @@
 ## Version 1.9.0 (Future)
 - Make date fields return carbon object
 - Add `loadRelationFields` method
+- Add abstract tags field
 
 ## Version 1.8.1 (2017-07-19)
 - Fix template creation from command line